# Save ~2800 bytes of RAM by not caching angle->trigger mapping