@racer-coder 2021-11-08
sorry this only saves 1440 bytes.  I thought it was double because of the way I was finding the large users.