@rusefillc 2021-11-09
I am hesitant because we are flying blind without an explicit performance test and explicit performance requirement

Can I pull the "what about" and create a couple of tickets for the skeletons I am aware of?