@mck1117 2021-11-09
> I am hesitant because we are flying blind without an explicit performance test

This netted less change than we can measure with perf tracing. Seriously.

Testing with an externally generated 60-2 at ~6k RPM.

Whole interrupt takes ~22us.  `updateInstantRpm` claims to take ~3us in either case, which is low enough that the perf trace overhead becomes significant.

TIME FOR THE LOGIC ANALYZER AND BIT TWIDDLING

I set/reset an output pin when we enter/leave `updateInstantRpm`.

Before this PR: 1.46 us
After this PR: 2.66 us

Adds ~1.2us, about 200 CPU cycles on an F4 at 168MHz.

And this was on a 60-2, just about our worst case common scenario for trigger size.  The cost grows with O(log(n)) of the trigger size, and in cases like this, O(log(n)) rounds to O(1).