@racer-coder 2021-11-09
Each iteration of the bisection was 11 instructions.  I'm not sure what the cycle count on them is.  I couldn't get it much lower without making the code very strange.  There are some weird things in the function, like calculating the # of buckets to check in floating point with a bunch of conditionals, only to cast it to an int later.  We could always cache that.

Also changing the lookup to ints would shave an instruction - floating point comparisons have one more instruction than integer comparisons.  Again, I'm not sure the cycle impact of that.

And we can hide the 1.2uS with #3495 :-)