@mck1117 2021-11-09
Not sure we need to cut down the cycle count any more.  11/iteration is plenty good, especially with the function still legible.  1us per trigger tooth is a tiny fish to fry at this point.  Especially since it saves ~1.5k of memory.